+++ projects/bot-api/src/main/java/com/z3rp436/discord/bot/api/persistence/BotTokenActiveRepository.java
@@ -1,0 +1,13 @@
+package com.z3rp436.discord.bot.api.persistence;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface BotTokenActiveRepository extends JpaRepository<BotTokenActiveEntity, Long> {
+
+    Optional<BotTokenActiveEntity> findByBotId(String botId);
+
+    void deleteByBotId(String botId);
+}
+